+++ package.json
@@ -37,6 +37,7 @@
     "rollup-plugin-livereload": "^2.0.5",
     "sass": "1.53.0",
     "siyuan": "1.0.8",
+    "siyuan-app": "git+https://github.com/siyuan-note/siyuan.git#v3.1.20",
     "svelte": "^4.2.19",
     "svelte-eslint-parser": "^0.41.1",
     "ts-node": "^10.9.1",
